requirements.txt
numpy~=1.26.4
ffmpeg-python~=0.2.0
pygame==2.5.2